+++ packages/device_info_plus/device_info_plus/android/src/main/java/dev/fluttercommunity/plus/device_info/DeviceInfoPlusPlugin.java
@@ -4,16 +4,23 @@
 
 package dev.fluttercommunity.plus.device_info;
 
+import android.app.Activity;
 import android.content.Context;
+
 import androidx.annotation.NonNull;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodChannel;
 
-/** DeviceInfoPlusPlugin */
-public class DeviceInfoPlusPlugin implements FlutterPlugin {
+/**
+ * DeviceInfoPlusPlugin
+ */
+public class DeviceInfoPlusPlugin implements FlutterPlugin, ActivityAware, IGetActivity {
 
   MethodChannel channel;
+  Activity activity;
 
   @Override
   public void onAttachedToEngine(FlutterPlugin.FlutterPluginBinding binding) {
@@ -25,10 +32,35 @@ public class DeviceInfoPlusPlugin implements FlutterPlugin {
     tearDownChannel();
   }
 
+  @Override
+  public void onAttachedToActivity(@NonNull ActivityPluginBinding binding) {
+    activity = binding.getActivity();
+  }
+
+  @Override
+  public void onDetachedFromActivityForConfigChanges() {
+    activity = null;
+  }
+
+  @Override
+  public void onReattachedToActivityForConfigChanges(@NonNull ActivityPluginBinding binding) {
+    activity = binding.getActivity();
+  }
+
+  @Override
+  public void onDetachedFromActivity() {
+    activity = null;
+  }
+
+  @Override
+  public Activity getActivity() {
+    return activity;
+  }
+
   private void setupMethodChannel(BinaryMessenger messenger, Context context) {
     channel = new MethodChannel(messenger, "dev.fluttercommunity.plus/device_info");
-    final MethodCallHandlerImpl handler =
-        new MethodCallHandlerImpl(context.getContentResolver(), context.getPackageManager());
+    final MethodCallHandlerImpl handler = new MethodCallHandlerImpl(context.getContentResolver(),
+      context.getPackageManager(), this);
     channel.setMethodCallHandler(handler);
   }