+++ packages/device_info_plus/device_info_plus/android/src/main/java/dev/fluttercommunity/plus/device_info/IGetActivity.java
@@ -1,0 +1,7 @@
+package dev.fluttercommunity.plus.device_info;
+
+import android.app.Activity;
+
+interface IGetActivity {
+	Activity getActivity();
+}